+++ Changelog.md
@@ -3,8 +3,10 @@
 - New: The Alarm, Motion Activated Light, and Portable Radar can now be waterlogged
 - New: Laser and inventory scanner fields can now be waterlogged, which means the Laser and Inventory Scanner now properly work underwater without air pockets
 - New: Randomize signature button for the Keycard Reader
+- New: The pitch of the sound an alarm plays can now be changed
 - Change: Several technical blocks' sounds have been adjusted to better match how they look
 - Change: Improved visuals when holding a taser
+- Change: Some alarm options have been moved to a separate screen
 - Fix: Shields do not take damage when blocking a taser that deals high enough damage
 - Fix: The death message of a player killed by a taser does not contain the player firing the taser
 - Fix: Crashes when rendering some modded block entities in a Projector, or using them as a disguise